+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - For the `show` command, either `--summary` or `--summaey-with-items` can be set but not both
+- Rename `Transportation` to `Travel` in the demo categories
 
 ## [0.5.2] - 2025-02-14